UK electricity load data

Dataset on UK electricity demand, taken from the national grid (http://www2.nationalgrid.com/).

data(UKload)

Format

UKload contains the following variables:

NetDemand

net electricity demand between 11:30am and 12am.

wM

instantaneous temperature, averaged over several English cities.

wM_s95

exponential smooth of wM, that is wM_s95[i] = a*wM[i] + (1-a)*wM_s95[i] with a=0.95

.
Posan

periodic index in [0, 1] indicating the position along the year.

Dow

factor variable indicating the day of the week.

Trend

progressive counter, useful for defining the long term trend.

NetDemand.48

lagged version of NetDemand, that is NetDemand.48[i] = NetDemand[i-2].

Holy

binary variable indicating holidays.

Year

should be obvious.

Date

should be obvious.

Details

See Fasiolo et al. (2017) for details.

Value

matrix of replicate data series

References

Fasiolo, M., Goude, Y., Nedellec, R. and Wood, S. N. (2017). Fast calibrated additive quantile regression. Available at https://arxiv.org/abs/1707.03307.

Examples

# NOT RUN {
  library(qgam)
  data(UKload)
  plot(UKload$NetDemand, type = 'l')
# }