Some diagnostics for a fitted qgam model

Takes a fitted gam object produced by qgam() and produces some diagnostic information about the fitting procedure and results. It is a simplified version of mgcv::gam.check.

# S3 method for qgam
check(obj, nbin = 10, lev = 0.05, ...)

Arguments

obj
the output of a qgam() call.

nbin
number of bins used in the internal call to cqcheck().

lev
the significance levels used by cqcheck(), which determines the width of the confidence intervals.

...
extra arguments to be passed to plot()

Value

Simply produces a plot and prints out some diagnostics.

Details

This function provides a single plot, which shows how the number of responses falling below the fitted quantile (y-axis) changes with the fitted quantile (x-axis). To be clear: if the quantile is fixed to, say, 0.5 we expect 50% of the responses to fall below the fit. See ?cqcheck() for details. The function also provide some convergence diagnostics (regarding the optimization), which are the same as in mgcv::gam.check. It reports also the maximum (k') and the selected degrees of freedom of each smooth term.

Examples

# NOT RUN {
library(qgam)
set.seed(0)
dat <- gamSim(1, n=200)
b<-qgam(y~s(x0)+s(x1)+s(x2)+s(x3), data=dat, qu = 0.5)
plot(b, pages=1)
check.qgam(b, pch=19, cex=.3)

# }